software/repo/services/worker/requirements.txt
celery==5.4.0
redis==5.0.3
psycopg[binary]==3.1.19
pydantic==2.7.4
boto3==1.34.144
python-dotenv==1.0.1
pyyaml==6.0.1
cryptography==41.0.7
